Add support for CustomJS.from_file()
